Commit of bug-fix in-progress Bug: width & height of projection window as related to image transformation.

+++ svl/include/cinder_cv/affine_rectangle.h
@@ -87,7 +87,7 @@ public:
     const cv::RotatedRect& rotatedRectInImage(const Area& image_bounds) const;
     
 private:
-    size_t getNearestIndex( const ivec2 &pt ) const;
+    int getNearestIndex( const ivec2 &pt ) const;
     
     void cornersInImage(const Area& image_bounds) const;
     mutable EditableRect   mRectangle;
@@ -99,7 +99,7 @@ private:
     bool           mIsOver;
     bool           mMouseIsDown;
     
-    size_t          mSelected;
+    int          mSelected;
     Rectf           mDisplayRect;
     Rectf           mPaddedRect;
     Rectf           mImageRect;
